+++ apps/nestjs-wallet/src/vc-api/exchanges/entities/transaction.entity.ts
@@ -1,6 +1,5 @@
 import { Column, Entity, JoinColumn, OneToOne } from 'typeorm';
 import { v4 as uuidv4 } from 'uuid';
-import { IPresentation, IPresentationDefinition, PEX, ProofPurpose } from '@sphereon/pex';
 import { ExchangeResponseDto } from '../dtos/exchange-response.dto';
 import { PresentationReviewStatus } from '../types/presentation-review-status';
 import { VerifiablePresentation } from '../../credentials/types/verifiable-presentation';
@@ -10,8 +9,7 @@ import { PresentationReviewEntity } from './presentation-review.entity';
 import { VpRequestEntity } from './vp-request.entity';
 import { CallbackConfiguration } from '../types/callback-configuration';
 import { PresentationSubmissionEntity } from './presentation-submission.entity';
-import { CredentialVerifier } from '../../credentials/types/credential-verifier';
-import { PresentationVerifier } from '../../credentials/types/presentation-verifier';
+import { VpRequestSubmissionVerifier } from 'src/vc-api/vp-request-submission-verifier';
 
 /**
  * A TypeOrm entity representing an exchange transaction
@@ -83,64 +81,6 @@ export class TransactionEntity {
   @Column('simple-json')
   callback: CallbackConfiguration[];
 
-  private verifyVpRequestTypeDidAuth(presentation: VerifiablePresentation): string[] {
-    // https://w3c-ccg.github.io/vp-request-spec/#did-authentication-request
-    const errors: string[] = [];
-
-    if (!presentation.holder) {
-      errors.push('Presentation holder is required for didAuth query');
-    }
-
-    return errors;
-  }
-
-  private verifyVpRequestTypePresentationDefinition(
-    presentation: VerifiablePresentation,
-    credentialQuery: Array<{ presentationDefinition: IPresentationDefinition }>
-  ): string[] {
-    // https://identity.foundation/presentation-exchange/#presentation-definition
-    const errors: string[] = [];
-    const pex: PEX = new PEX();
-
-    credentialQuery.forEach(({ presentationDefinition }, index) => {
-      const { errors: partialErrors } = pex.evaluatePresentation(
-        presentationDefinition,
-        presentation as IPresentation
-      );
-
-      errors.push(
-        ...partialErrors.map(
-          (error) =>
-            `Presentation definition (${index + 1}) validation failed, reason: ${error.message || 'Unknown'}`
-        )
-      );
-    });
-
-    return errors;
-  }
-
-  private validatePresentation(presentation: VerifiablePresentation): string[] {
-    const commonErrors = [];
-    // Common checking
-    if (presentation.proof.challenge !== this.vpRequest.challenge) {
-      commonErrors.push('Challenge does not match');
-    }
-
-    // Type specific checking
-    const partialErrors = this.vpRequest.query.flatMap((vpQuery) => {
-      switch (vpQuery.type) {
-        case VpRequestQueryType.didAuth:
-          return this.verifyVpRequestTypeDidAuth(presentation);
-        case VpRequestQueryType.presentationDefinition:
-          return this.verifyVpRequestTypePresentationDefinition(presentation, vpQuery.credentialQuery);
-        default:
-          return ['Unknown request query type'];
-      }
-    });
-
-    return [...partialErrors, ...commonErrors];
-  }
-
   /**
    * Process a presentation submission.
    * Check the correctness of the presentation against the VP Request Credential Queries.
@@ -149,32 +89,12 @@ export class TransactionEntity {
    */
   public async processPresentation(
     vp: VerifiablePresentation,
-    credentialVerifier: CredentialVerifier,
-    presentationVerifier: PresentationVerifier
+    verifier: VpRequestSubmissionVerifier
   ): Promise<{
     response: ExchangeResponseDto;
     callback: CallbackConfiguration[];
   }> {
-    const verifyOptions = {
-      challenge: this.vpRequest.challenge,
-      proofPurpose: ProofPurpose.authentication,
-      verificationMethod: vp.proof.verificationMethod as string //TODO: fix types here
-    };
-    const result = await presentationVerifier.verifyPresentation(vp, verifyOptions);
-    if (!result.checks.includes('proof') || result.errors.length > 0) {
-      return {
-        response: {
-          errors: [
-            `${this.transactionId}: verification of presentation proof not successful`,
-            ...result.errors
-          ]
-        },
-        callback: []
-      };
-    }
-
-    const errors = this.validatePresentation(vp);
-
+    const errors = await verifier.verifyPresention(vp, this.vpRequest);
     if (errors.length > 0) {
       return {
         response: {
